chore: Update folder structure
Changed folder structure to better align with SoC and Colocation
