+++ src/modules/home/Home.tsx
@@ -1,0 +1,92 @@
+import React from 'react';
+import { Inter } from '@next/font/google';
+import styles from './Home.module.css';
+
+const inter = Inter({ subsets: ['latin'] });
+
+export const Home: React.FC = () => {
+    return (
+        <main className={styles.main}>
+            <div className={styles.description}>
+                <p>
+                    Get started by editing&nbsp;
+                    <code className={styles.code}>pages/index.tsx</code>
+                </p>
+                <div>
+                    <a
+                        href="https://vercel.com?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
+                        target="_blank"
+                        rel="noopener noreferrer"
+                    >
+                        By <img src="/vercel.svg" alt="Vercel Logo" className={styles.vercelLogo} />
+                    </a>
+                </div>
+            </div>
+
+            <div className={styles.center}>
+                <img className={styles.logo} src="/next.svg" alt="Next.js Logo" />
+                <div className={styles.thirteen}>
+                    <img src="/thirteen.svg" alt="13" />
+                </div>
+            </div>
+
+            <div className={styles.grid}>
+                <a
+                    href="https://nextjs.org/docs?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
+                    className={styles.card}
+                    target="_blank"
+                    rel="noopener noreferrer"
+                >
+                    <h2 className={inter.className}>
+                        Docs <span>-&gt;</span>
+                    </h2>
+                    <p className={inter.className}>
+                        Find in-depth information about Next.js features and&nbsp;API.
+                    </p>
+                </a>
+
+                <a
+                    href="https://nextjs.org/learn?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
+                    className={styles.card}
+                    target="_blank"
+                    rel="noopener noreferrer"
+                >
+                    <h2 className={inter.className}>
+                        Learn <span>-&gt;</span>
+                    </h2>
+                    <p className={inter.className}>
+                        Learn about Next.js in an interactive course with&nbsp;quizzes!
+                    </p>
+                </a>
+
+                <a
+                    href="https://vercel.com/templates?framework=next.js&utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
+                    className={styles.card}
+                    target="_blank"
+                    rel="noopener noreferrer"
+                >
+                    <h2 className={inter.className}>
+                        Templates <span>-&gt;</span>
+                    </h2>
+                    <p className={inter.className}>
+                        Discover and deploy boilerplate example Next.js&nbsp;projects.
+                    </p>
+                </a>
+
+                <a
+                    href="https://vercel.com/new?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
+                    className={styles.card}
+                    target="_blank"
+                    rel="noopener noreferrer"
+                >
+                    <h2 className={inter.className}>
+                        Deploy <span>-&gt;</span>
+                    </h2>
+                    <p className={inter.className}>
+                        Instantly deploy your Next.js site to a shareable URL with&nbsp;Vercel.
+                    </p>
+                </a>
+            </div>
+        </main>
+    );
+};
