+++ src/modules/home/index.ts
@@ -1,0 +1,1 @@
+export { Home } from './Home';
