+++ src/pages/_app.tsx
@@ -1,5 +1,5 @@
 import 'normalize.css/normalize.css';
-import '@/styles/globals.css';
+import '@styles/globals.css';
 import type { AppProps } from 'next/app';
 import { DefaultSeo } from 'next-seo';
 import SEO from '../../next-seo.config';
